+++ workout-tracker/src/css/NavigationMenu.css
@@ -1,0 +1,36 @@
+.Navbar {
+   
+    display: flex;
+    width: 100%;
+    
+    height: 100%;
+    border: 100px;
+    
+    border-color: #ad0000;
+    color: #000000;
+    align-items: center;
+    justify-content: center;
+    
+}
+.menuButton {
+    display: inline;
+    min-width: fit-content;
+    min-height: fit-content;
+    width: 10%;
+    height: 50px;
+    background-color: rgb(110, 110, 110);
+    border-radius: 15%;
+    border: 10px;
+    margin: 2px;
+    color: #ffffff;
+    
+
+}
+.menuButton:hover{    
+    background-color: rgb(190, 190, 190);
+    color: #000;
+}
+.menuButton:active{
+    transform: translateY(1.5px);
+    box-shadow: -1px -1px #000000;
+}
